refactor(taggit): Renamed to tag_fields #1
taggit has been renamed and all the necessary changes downstream have
been updated and passed all the tests.
