+++ tag_fields/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from taggit.models import Tag, TaggedItem
+from tag_fields.models import Tag, TaggedItem
 
 
 class TaggedItemInline(admin.StackedInline):
