+++ tag_fields/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from django.utils.translation import gettext as _
 
-from taggit.utils import edit_string_for_tags, parse_tags
+from tag_fields.utils import edit_string_for_tags, parse_tags
 
 
 class TagWidgetMixin:
