+++ tag_fields/locale/es/LC_MESSAGES/django.po
