+++ tag_fields/locale/fa/LC_MESSAGES/django.po
