+++ tag_fields/locale/fr/LC_MESSAGES/django.mo
