+++ tag_fields/locale/nb/LC_MESSAGES/django.mo
