+++ tag_fields/locale/uk/LC_MESSAGES/django.mo
