+++ tag_fields/views.py
@@ -2,7 +2,7 @@ from django.contrib.contenttypes.models import ContentType
 from django.shortcuts import get_object_or_404
 from django.views.generic.list import ListView
 
-from taggit.models import Tag, TaggedItem
+from tag_fields.models import Tag, TaggedItem
 
 
 def tagged_object_list(request, slug, queryset, **kwargs):
